+++ changelog.md
@@ -12,6 +12,14 @@
 - Se agregó la opción de pausar el contrato.
 - Se agregaron los test de covertura.
 
+## [NFT v2.2.1] - 24-07-2023
+
+- Se eliminó opción pausable.
+- Se agregó función para mintear token de renta como comprobante.
+- Se agregó función para hacer burn a token de renta.
+- Valor isRented en struct de Car.
+- Setteo de valor isRented mediante función.
+
 ## [RentCar v2.0.0] - 20-07-2023
 
 - Versión inicial del contrato RentCarv2 con algunas correcciones del contrato base RentCar.
@@ -21,3 +29,11 @@
 - Se agregó la función setCommissionPercentage
 - Se eliminó el rol RENTER_ROL
 - Se agregó la función de versión para el contrato.
+
+## [RentCar v2.1.1] - 24-07-2023
+
+- Refactorización de contrato.
+- Se agregaron modificadores.
+- Setteo de comisión mediante función.
+- Mejora en funciones de cálculo para renta.
+- Se agregaron test iniciales de covertura.